Change iDocFrequency to keep precision on division
Title

+++ WI/ExerciseSession2/WI120917/WI120917/Program.cs
@@ -91,7 +91,7 @@ namespace WI120917
 
                 double weightedTermFrequency = 1 + Math.Log10(tf.Value);
                 int docFrequency = index[tf.Key].Count;
-                double iDocFrequency = Math.Log10(docCount / docFrequency);
+                double iDocFrequency = Math.Log10(docCount / (double)docFrequency);
                 double wt = weightedTermFrequency * iDocFrequency;
 
                 queryVector.Add(wt);
